% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{betas}
\alias{betas}
\title{Illumina 450K experimental data}
\format{A matrix with beta values:
\describe{
  \item{Rows}{CpG}
  \item{Columns}{Samples}
  ...
}}
\source{
\url{http://www.diamondse.info/}
}
\usage{
betas
}
\description{
A dataset containing beta values for 2 experimental condition in triplicats.
}
\keyword{datasets}

