% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAT.R
\name{DAT}
\alias{DAT}
\title{Compute averaged beta values and classify CpG
\code{DAT} performs the Double Average differencial methylation analysis.}
\usage{
DAT(betaValues, expDat, window.width = 3, threads = 1)
}
\arguments{
\item{betaValues}{A matrix containing beta values for each CpG (row) and sample (col) as can be obtained using the minfi::getBeta() package function.}

\item{expDat}{A dataframe containing experimental information about the samples as can be obtained from the read.450k() function in the minfi package}

\item{window.width}{Size of the window during the rolling average computation.}

\item{threads}{Number of threads to use for parallel computing.}
}
\value{
A list containing \enumerate{
     \item Names of samples in each condition
     \item Classification of CpG (named character vector)
     \item Beta values differences between condition (named numeric vector)
     \item Summary statistics
     }
}
\description{
Compute averaged beta values and classify CpG
\code{DAT} performs the Double Average differencial methylation analysis.
}
\examples{
DAT(betas[1:10000,],expSamples,threads=10)
}

