% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{expSamples}
\alias{expSamples}
\title{Illumina 450K experimental setup}
\format{A dataframe describing samples as in the Minfi Package:
\describe{
   \item{Sample_Name}{names of samples}
   \item{Sample_Group}{Experimental condition, must be a factor with first level corresponding to the control}
   \item{Pool_ID}{Index for paired analysis}
   }}
\source{
\url{http://www.diamondse.info/}
}
\usage{
expSamples
}
\description{
A dataset containing beta values for 2 experimental condition in triplicats.
}
\keyword{datasets}

